Fix file cleanup logic

+++ destroy.sh
@@ -1,14 +1,19 @@
 #!/bin/bash
 
-env USER_GID=$(id -g) USER_UID=$(id -u) docker compose down -v
+docker compose down -v
 
-echo -ne "Do you want to remove the following files? (y|n) \n"
 git clean -fdxn
-read option
-if [ "$option" == "y" ]; then
-    git clean -fdx
-elif [ "$option" == "n" ]; then
-    echo -e "\nNot removing any untracked files.\n"
-else 
-    echo "Sorry, I could not understand the input, so no files were removed."
+if [ ! -n $? ]; then
+    echo -ne "Do you want to remove the following files? (y|n) \n"
+    read option
+    if [ "$option" == "y" ]; then
+        git clean -fdx
+    elif [ "$option" == "n" ]; then
+        echo -e "\nNot removing any untracked files.\n"
+    else 
+        echo -e "\nSorry, I could not understand the input.\n"
+    fi
+else
+    echo -e "\nNo untracked files were found.\n"
+    exit 0
 fi
